build(package): generate package using top level root as-is
