--- .npmignore
@@ -1,7 +1,0 @@
-perf
-spec
-.git
-yarn.lock
-coverage
-typings/
-docs_app
