+++ package.json
@@ -1,8 +1,11 @@
 {
-  "name": "@reactivex/rxjs",
+  "name": "rxjs",
   "version": "7.0.0-alpha.0",
   "description": "Reactive Extensions for modern JavaScript",
-  "main": "index.js",
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/esm5/index.js",
+  "es2015": "./dist/esm/index.js",
+  "types": "./dist/types/index.d.ts",
   "sideEffects": false,
   "config": {
     "commitizen": {
@@ -22,42 +25,13 @@
   "scripts": {
     "precommit": "lint-staged",
     "commitmsg": "validate-commit-msg",
-    "build_all": "npm-run-all clean_dist copy_sources build_cjs build_esm5 build_esm2015 build_esm5_for_rollup build_umd build_types generate_packages",
-    "build_cjs": "npm-run-all clean_dist_cjs compile_dist_cjs",
-    "build_esm5": "npm-run-all clean_dist_esm5 compile_dist_esm5",
-    "build_esm5_for_rollup": "npm-run-all clean_dist_esm5_for_rollup compile_dist_esm5_for_rollup && mkdirp dist/esm5_for_rollup/node_modules",
-    "build_esm2015": "npm-run-all clean_dist_esm2015 compile_dist_esm2015",
-    "build_types": "npm-run-all clean_dist_types compile_dist_types",
-    "build_closure_core": "node ./tools/make-closure-core.js",
-    "build_global": "npm-run-all clean_dist_global build_esm5_for_rollup && mkdirp ./dist/global && node ./tools/make-umd-bundle.js && npm-run-all build_closure_core clean_dist_esm5_for_rollup",
-    "build_umd": "npm-run-all clean_dist_global && mkdirp ./dist/global && node ./tools/make-umd-bundle.js && npm-run-all build_closure_core",
-    "build_perf": "webdriver-manager update && npm-run-all build_cjs build_global perf",
-    "build_spec": "npm-run-all build_cjs generate_packages",
-    "build_spec_full": "npm-run-all build_spec",
     "build:spec:browser": "echo \"Browser test is not working currently\" && exit -1 && webpack --config spec/support/webpack.mocha.config.js",
-    "clean_dist": "shx rm -rf ./dist",
-    "clean_dist_cjs": "shx rm -rf ./dist/cjs",
-    "clean_dist_esm5": "shx rm -rf ./dist/esm5",
-    "clean_dist_esm5_for_rollup": "shx rm -rf ./dist/esm5_for_rollup",
-    "clean_dist_esm2015": "shx rm -rf ./dist/esm2015",
-    "clean_dist_global": "shx rm -rf ./dist/global",
-    "clean_dist_types": "shx rm -rf ./dist/typings",
-    "commit": "git-cz",
-    "compile_dist_cjs": "tsc -p ./tsconfig/tsconfig.cjs.json",
-    "compile_dist_esm5": "tsc -p ./tsconfig/tsconfig.esm5.json",
-    "compile_dist_esm2015": "tsc -p ./tsconfig/tsconfig.esm2015.json",
-    "compile_dist_esm5_for_rollup": "tsc -p ./tsconfig/tsconfig.esm5.rollup.json",
-    "compile_dist_types": "tsc -p ./tsconfig/tsconfig.types.json",
-    "copy_sources": "mkdirp dist && shx cp -r ./src/ ./dist/src",
     "decision_tree_widget": "cd doc/decision-tree-widget && npm run build && cd ../..",
-    "generate_packages": "node .make-packages.js",
-    "lint_perf": "eslint perf/",
     "lint_spec": "tslint -c tslint.json \"spec/**/*.ts\"",
     "lint_src": "tslint -c tslint.json \"src/**/*.ts\"",
     "lint": "npm-run-all --parallel lint_*",
-    "dtslint": "npm run build_types && dtslint --localTs ./node_modules/typescript/lib --expectOnly ./spec-dtslint",
-    "perf:micro": "node ./perf/micro/index.js",
-    "prepublish": "shx rm -rf ./typings && npm run build_all",
+    "dtslint": "npm run compile && dtslint --localTs ./node_modules/typescript/lib --expectOnly ./spec-dtslint",
+    "prepublishOnly": "npm run build:package",
     "postpublish": "./docs_app/scripts/publish-docs.sh",
     "publish_docs": "./publish_docs.sh",
     "test": "cross-env TS_NODE_PROJECT=spec/tsconfig.json mocha --opts spec/support/default.opts \"spec/**/*-spec.ts\"",
@@ -68,7 +42,11 @@
     "test:systemjs": "node integration/systemjs/systemjs-compatibility-spec.js",
     "test:side-effects": "check-side-effects --test integration/side-effects/side-effects.json",
     "test:side-effects:update": "npm run test:side-effects -- --update",
-    "tests2png": "mkdirp docs_app/content/img && mocha --opts spec/support/tests2png.opts \"spec/**/*-spec.ts\""
+    "tests2png": "mkdirp docs_app/content/img && mocha --opts spec/support/tests2png.opts \"spec/**/*-spec.ts\"",
+    "compile": "tsc -b --verbose ./src/tsconfig.cjs.json ./src/tsconfig.esm.json ./src/tsconfig.esm5.json ./src/tsconfig.esm5.rollup.json ./src/tsconfig.types.json ./spec/tsconfig.json",
+    "build:clean": "shx rm -rf ./dist",
+    "build:global": "node ./tools/make-umd-bundle.js && node ./tools/make-closure-core.js",
+    "build:package": "npm-run-all build:clean compile build:global && node ./tools/prepare-package.js && node ./tools/generate-alias.js"
   },
   "repository": {
     "type": "git",
@@ -180,8 +158,23 @@
     "validate-commit-msg": "2.14.0",
     "webpack": "^4.31.0"
   },
-  "engines": {
-    "npm": ">=2.0.0"
-  },
-  "typings": "./dist/package/Rx.d.ts"
+  "files": [
+    "dist/bundles",
+    "dist/cjs/**/!(*.tsbuildinfo)",
+    "dist/esm/**/!(*.tsbuildinfo)",
+    "dist/esm5/**/!(*.tsbuildinfo)",
+    "dist/types/**/!(*.tsbuildinfo)",
+    "ajax",
+    "fetch",
+    "operators",
+    "testing",
+    "webSocket",
+    "src",
+    "CHANGELOG.md",
+    "CODE_OF_CONDUCT.md",
+    "LICENSE.txt",
+    "package.json",
+    "README.md",
+    "tsconfig.json"
+  ]
 }
